refactor(WidgetDataStore): extract QuotaNotificationEvaluator
Pulls the quota pace-notification subsystem out of WidgetDataStore into a
focused QuotaNotificationEvaluator: the per-metric dedup state, the
fire/deliver/commit decision, the notification build, and the ~70 lines of
debug trace helpers. WidgetDataStore now just gathers each pass's enabled,
bounded, visible metrics and delegates; delivery + provider display name pass
in as closures so the evaluator is independent of the store's providers.

`evaluateNotifications(now:)` stays on the store (unchanged signature — the
notification tests call it), so this is purely internal restructuring.
WidgetDataStore drops from 634 to 502 LOC (at the ~500 rule), leaving it
focused on refresh orchestration + MetricLine resolution.

Verified: swift build + 883 tests green (WidgetDataStoreNotificationTests
exercise the fire/deliver/dedup/prune paths end-to-end through the evaluator).
